Move test locations
